requirements.txt
mitmproxy==11.0.0
stem==1.8.2
python-dotenv==1.0.1